Add a link to show user subpages for each student in the articles table.
Add an additional link, intended to make it easy to find student
sandboxes, to the set of links that appears after each username
in the table of students and articles on each course page, by
changing the getToolLinks functions. Also change labels for
courses that have ended from Passed to Previous to be less
confusable with the concept of students having "passed" a course.

Change-Id: Ia4fe883842bdc998ad997fe0b148f1f8b0ff6ef5