Fix errors if course/org forms are re-submitted
Provides one of two changes suggested for bug 43782
(duplicate entries for the same course). If the user creates a
course and then clicks the "Submit" button a second time, a duplicate
course is no longer created. This was possible if the user clicked the
browser's back button after creating a course. It may also have
occurred when the network was iffy. In the case of institutions,
instead of a duplicate entry, the user would an internal error.
This patch fixes the issue for both courses and institutions.

Bug: 43782
Change-Id: Ia672d955e0995c5593c53aa7e0923178418b3915