Fix student name in dialogue for removing article
In the students table at the end of course pages, next to articles
assigned to students, there are links for removing article
assignments. When you click any of those links, a confirmation
dialogue appears. If you click a link to remove an article for a
student other than yourself, the message in the dialogue shows your
user name instead of the student's. This little patch fixes the error.

Change-Id: I6b89544cfa4cdc08aab4ac423074198c64dac85a