Disable institution-page cache when logged in
The caching of institution pages may confuse users, since changes they make
using controls on this page may not show up right away, and users may not
notice or understand the message saying that they're viewing a cached version
of the page. There is no need to change the behavior for anonymous users, since
said controls are only available to logged-in users with certain privileges. It
seems unlikely that this change will cause performance issues.

Change-Id: I8ed3e9fc1e16dc4da58d02f2596560262c546e25