On Special:Contributions, mention EP roles
The Special:Contributions page gives you an idea of what a user is
up to. To understand and collaborate with a user, it may be useful
to know that his or her contributions were course assignments, or
that he or she is an instructor or volunteer for Education Program
courses on Wikipedia. However, it can be difficult to identify
contributions related to the EP, and it will be even harder to do
so once change I9f8db940b841607ba3f6f99441c5c8a3dd662e35 is merged.
This patch addresses the issue by adding a message and some links
to the top of the Special:Contributions page for users with
Education Program roles.

Change-Id: I409b9ea1d0165bc6d73afe3cd4f44d3b52bc2f2d