Allow for URLs in article assignment
Users could potentially misuse the article assignment box by inputting
a URL instead of the title of an article. This bugfix parses the input
and tries to add the article referenced by the end of the URL, but
only for like-domained URLs.
(e.g. http://en.wikipedia.org/wiki/Greece -> Greece)

Bug: 44397
Change-Id: I294835999c265d309af05c3de9869ae353a10ea7