Delay check of cldr presence after Setup
To test extensions, Jenkins clones the repositories and then load them
using scandir().  That can cause EducationProgram to be loaded before
cldr which will early abort the execution since the die() is in the
entry point.

Instead, rely on the SetupAfterCache hook which is triggered after
LocalSettings.php.  Thus if cldr has been required, EducationProgram
should be just fine.

Bug: 67598
Change-Id: I592d9b31dbf5b37520e6212d7d6160497c15d4a2