Never unconditionally 'return false' from interface hooks
Reverts e9a1d54.

This was causing the VisualEditor "Edit" tab to be missing on
en.wikipedia.org in all pages. It worked fine on test.wikipedia.org
because EducationProgram is not installed there.

At run-time, JavaScript fixes this up (because VE supports live configuation
changes to counter-act static page caching when e.g. VE is enabled
in additional namespaces) and that code was covering this bug.

However due to code now loading asynchronously in VE, the fact
that this is added at run-time is now a lot more visible and causing
flashes.

Bug: T107980
Change-Id: I3105f285238ba6b10fdce22198c7dc5e2d13b588