Ensure that deletion messages are not unescaped in the error dialog
The error messages that can be displayed when someone tries to delete
an institution are plain text and correctly escaped when output into
the data-no-del-text attribute. However, when someone tries to bulk
delete an institution and either doesn't have permission to do so or
if the institution has courses, the error message is retrieved from
the attribute unescaped and then put into a dialog as raw HTML. This
means that the MediaWiki system messages for these errors can be used
to inject raw HTML.

This fixes the issue by ensuring the messages are escaped when they
are put into the content of the error dialog.

Change-Id: I31e69e3285b88757c24b1a184874cd4ff0518dea